+++ windows_service/tests/test_windows_service.py
@@ -1,9 +1,10 @@
 # (C) Datadog, Inc. 2018-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import ctypes
+
 import pytest
 import pywintypes
-import ctypes
 import winerror
 from mock import patch
 from six import PY2
@@ -257,6 +258,7 @@ def test_invalid_pattern_regex(aggregator, check, instance_basic_dict):
     with pytest.raises(Exception, match=r"Regular expression syntax error in '\(foo':"):
         c.check(instance_basic_dict)
 
+
 def test_trigger_start(aggregator, check, instance_trigger_start):
     c = check(instance_trigger_start)
     c.check(instance_trigger_start)
@@ -274,16 +276,21 @@ def test_trigger_start(aggregator, check, instance_trigger_start):
         count=1,
     )
 
+
 def test_trigger_count_failure(aggregator, check, instance_trigger_start, caplog):
     c = check(instance_trigger_start)
 
-    with patch('datadog_checks.windows_service.windows_service.QueryServiceConfig2W', side_effect=[ctypes.WinError(winerror.ERROR_INSUFFICIENT_BUFFER), ctypes.WinError(1)]*2):
+    with patch(
+        'datadog_checks.windows_service.windows_service.QueryServiceConfig2W',
+        side_effect=[ctypes.WinError(winerror.ERROR_INSUFFICIENT_BUFFER), ctypes.WinError(1)] * 2,
+    ):
         c.check(instance_trigger_start)
-    
+
     if PY2:
         assert 'WindowsError: [Error 1] Incorrect function' in caplog.text
     else:
-        assert 'OSError: [WinError 1] Incorrect function' in caplog.text                                                                                              
+        assert 'OSError: [WinError 1] Incorrect function' in caplog.text
+
 
 @pytest.mark.e2e
 def test_basic_e2e(dd_agent_check, check, instance_basic):